feat: ajout docker-compose + corrections Layered + init projet Hexagonal

+++ layered-architecture/src/main/java/com/architecturelab/layered/service/ProductService.java
@@ -1,42 +1,58 @@
 package com.architecturelab.layered.service;
 
+import com.architecturelab.layered.dto.ProductDTO;
+import com.architecturelab.layered.dto.ProductMapper;
 import com.architecturelab.layered.exception.ProductNotFoundException;
 import com.architecturelab.layered.model.Product;
 import com.architecturelab.layered.repository.ProductRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ProductService {
+
     private final ProductRepository repository;
+    private final ProductMapper mapper;
 
-    public ProductService(ProductRepository repository) {
-        this.repository = repository;
+    // Récupérer tous les produits en DTO
+    public List<ProductDTO> findAll() {
+        return mapper.toDtoList(repository.findAll());
     }
 
-    public List<Product> findAll() {
-        return repository.findAll();
+    // Récupérer par id
+    public ProductDTO findById(Long id) {
+        Product product = repository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException(id));
+        return mapper.toDto(product);
     }
 
-    public Product findById(Long id) {
-      //  return repository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
-        return repository.findById(id).orElseThrow(() -> new ProductNotFoundException(id));
-
+    // Créer un produit
+    public ProductDTO create(ProductDTO dto) {
+        Product product = mapper.toEntity(dto);
+        Product saved = repository.save(product);
+        return mapper.toDto(saved);
     }
 
-    public Product create(Product product) {
-        return repository.save(product);
+    // Mettre à jour
+    public ProductDTO update(Long id, ProductDTO dto) {
+        Product existing = repository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException(id));
+
+        existing.setName(dto.getName());
+        existing.setPrice(dto.getPrice());
+
+        Product updated = repository.save(existing);
+        return mapper.toDto(updated);
     }
 
-    public Product update(Long id, Product updated) {
-        Product existing = findById(id);
-        existing.setName(updated.getName());
-        existing.setPrice(updated.getPrice());
-        return repository.save(existing);
-    }
-
+    // Supprimer
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ProductNotFoundException(id);
+        }
         repository.deleteById(id);
     }
 }
